feat: extend cafs object
