+++ packages/cafs-types/README.md
@@ -1,0 +1,28 @@
+# @pnpm/types
+
+> Types for the cafs
+
+<!--@shields('npm')-->
+[![npm version](https://img.shields.io/npm/v/@pnpm/types.svg)](https://www.npmjs.com/package/@pnpm/types)
+<!--/@-->
+
+## Installation
+
+```sh
+pnpm add @pnpm/types
+```
+
+## Usage
+
+```ts
+import { PackageManifest } from '@pnpm/types'
+
+const pkg: PackageManifest = {
+  name: 'foo',
+  version: '1.0.0',
+}
+```
+
+## License
+
+MIT
