+++ packages/cafs-types/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@pnpm/store-controller-types",
-  "version": "14.1.0",
-  "description": "Types for the store controller",
+  "name": "@pnpm/cafs-types",
+  "version": "0.0.0",
+  "description": "Types for the cafs",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "engines": {
@@ -17,7 +17,7 @@
     "compile": "tsc --build && pnpm run lint --fix",
     "lint": "eslint src/**/*.ts"
   },
-  "repository": "https://github.com/pnpm/pnpm/blob/main/packages/store-controller-types",
+  "repository": "https://github.com/pnpm/pnpm/blob/main/packages/cafs-types",
   "keywords": [
     "pnpm7",
     "pnpm",
@@ -27,17 +27,13 @@
   "bugs": {
     "url": "https://github.com/pnpm/pnpm/issues"
   },
-  "homepage": "https://github.com/pnpm/pnpm/blob/main/packages/store-controller-types#readme",
-  "dependencies": {
-    "@pnpm/fetcher-base": "workspace:*",
-    "@pnpm/resolver-base": "workspace:*",
-    "@pnpm/types": "workspace:*"
-  },
-  "funding": "https://opencollective.com/pnpm",
   "devDependencies": {
     "@pnpm/cafs-types": "workspace:*",
-    "@pnpm/store-controller-types": "workspace:*"
+    "@pnpm/types": "workspace:*",
+    "@types/ssri": "^7.1.1"
   },
+  "homepage": "https://github.com/pnpm/pnpm/blob/main/packages/cafs-types#readme",
+  "funding": "https://opencollective.com/pnpm",
   "exports": {
     ".": "./lib/index.js"
   }
