+++ packages/create-cafs-store/tsconfig.json
@@ -16,6 +16,9 @@
       "path": "../cafs"
     },
     {
+      "path": "../cafs-types"
+    },
+    {
       "path": "../fetcher-base"
     },
     {
