+++ packages/fetcher-base/src/index.ts
@@ -1,55 +1,11 @@
 import { Resolution, GitResolution, DirectoryResolution } from '@pnpm/resolver-base'
-import { DependencyManifest } from '@pnpm/types'
-import { IntegrityLike } from 'ssri'
-
-export interface PackageFileInfo {
-  checkedAt?: number // Nullable for backward compatibility
-  integrity: string
-  mode: number
-  size: number
-}
-
-export type PackageFilesResponse = {
-  fromStore: boolean
-  packageImportMethod?: 'auto' | 'hardlink' | 'copy' | 'clone' | 'clone-or-copy'
-  sideEffects?: Record<string, Record<string, PackageFileInfo>>
-} & ({
-  local: true
-  filesIndex: Record<string, string>
-} | {
-  local?: false
-  filesIndex: Record<string, PackageFileInfo>
-})
-
-export interface ImportPackageOpts {
-  requiresBuild?: boolean
-  sideEffectsCacheKey?: string
-  filesResponse: PackageFilesResponse
-  force: boolean
-}
-
-export type ImportPackageFunction = (
-  to: string,
-  opts: ImportPackageOpts
-) => Promise<{ isBuilt: boolean, importMethod: undefined | string }>
-
-export interface Cafs {
-  addFilesFromDir: (dir: string, manifest?: DeferredManifestPromise) => Promise<FilesIndex>
-  addFilesFromTarball: (stream: NodeJS.ReadableStream, manifest?: DeferredManifestPromise) => Promise<FilesIndex>
-  importPackage: ImportPackageFunction
-  tempDir: () => Promise<string>
-}
+import type { DeferredManifestPromise, Cafs, FilesIndex } from '@pnpm/cafs-types'
 
 export interface FetchOptions {
   manifest?: DeferredManifestPromise
   lockfileDir: string
   onStart?: (totalSize: number | null, attempt: number) => void
   onProgress?: (downloaded: number) => void
-}
-
-export interface DeferredManifestPromise {
-  resolve: (manifest: DependencyManifest | undefined) => void
-  reject: (err: Error) => void
 }
 
 export type FetchFunction<FetcherResolution = Resolution, Options = FetchOptions, Result = FetchResult> = (
@@ -64,19 +20,6 @@ export type FetchResult = {
 } | {
   local: true
   filesIndex: Record<string, string>
-}
-
-export interface FileWriteResult {
-  checkedAt: number
-  integrity: IntegrityLike
-}
-
-export interface FilesIndex {
-  [filename: string]: {
-    mode: number
-    size: number
-    writeResult: Promise<FileWriteResult>
-  }
 }
 
 export interface GitFetcherOptions {
