+++ packages/fetcher-base/tsconfig.json
@@ -10,6 +10,9 @@
   ],
   "references": [
     {
+      "path": "../cafs-types"
+    },
+    {
       "path": "../resolver-base"
     },
     {
