+++ packages/node.fetcher/tsconfig.json
@@ -13,6 +13,9 @@
       "path": "../../privatePackages/prepare"
     },
     {
+      "path": "../cafs-types"
+    },
+    {
       "path": "../create-cafs-store"
     },
     {
