+++ packages/package-requester/package.json
@@ -63,6 +63,7 @@
     "ssri": "^8.0.1"
   },
   "devDependencies": {
+    "@pnpm/cafs-types": "workspace:*",
     "@pnpm/client": "workspace:*",
     "@pnpm/create-cafs-store": "workspace:*",
     "@pnpm/logger": "^4.0.0",
