+++ packages/package-requester/src/index.ts
@@ -2,5 +2,4 @@ import packageRequester from './packageRequester'
 
 export default packageRequester
 
-export { PackageFilesResponse } from '@pnpm/fetcher-base'
 export { PackageResponse } from '@pnpm/store-controller-types'
