+++ packages/package-requester/tsconfig.json
@@ -16,6 +16,9 @@
       "path": "../cafs"
     },
     {
+      "path": "../cafs-types"
+    },
+    {
       "path": "../client"
     },
     {
